Make the carrier score the tenant's own, and stop leaking it
carrier_score_snapshots was declared a platform-wide table: no
tenant_id, no RLS, unique on carrier + scope + period + formula version.
The observations behind it come from shipments and delivery_outcomes,
which are under FORCE ROW LEVEL SECURITY, so the application role reads
only its own tenant. The scheduled recalculation walks tenants one by
one and overwrote that single row — leaving whichever tenant ran last,
and serving it to everyone through GET /v1/analytics/carriers. On-time
rate, incident rate, price index and sample size of one client were
presented to another as properties of the carrier: sample size is a
neighbour's shipment volume, price index the level of their tariffs.

The decision is recorded in ADR-0017 and is the human's: the score is
per tenant. Migration 0009 adds tenant_id, RLS with FORCE and the same
tenant_isolation policy the other business tables carry, and puts
tenant_id first in the uniqueness key so one tenant's recalculation can
no longer overwrite another's. Existing rows are deleted rather than
assigned: nobody knows whose observations produced them, and handing
them to a tenant would pass off someone else's numbers as theirs one
more time. The next recalculation rebuilds them; until then the screen
says there is not enough data, which is more honest than a number that
belongs to someone else.

Three tests cover it, each verified against a database restored to the
old shape — permissive policy and the old uniqueness key — where all
three fail: another tenant sees no snapshot, recalculating one tenant
does not move another's numbers, and RLS refuses a write addressed to
someone else's tenant.

With that settled, FR-7.6 is unblocked and done: score_at_quote,
score_confidence and score_scope are filled when the quote is taken and
stay in the offer as a snapshot, so a decision is still explained by the
numbers that were visible at the time — a test seeds two hundred bad
deliveries, recalculates, and asserts the recorded score did not move.
insufficient becomes null in the response: the RateOffer schema has no
such value, and with no number there is nothing to be confident about.

Two things found on the way. carrier_score left as a JSON string because
the field was a Decimal while the contract says number; the column is an
integer and the field now is too. And test_city_mapping never truncated,
so it passed only because the files that commit a carrier with the same
code happened to run after it; its fixture now asks for clean_db instead
of relying on alphabetical order.

603 backend tests, ruff, mypy, seven import-linter contracts, alembic
check, a full downgrade and upgrade of the new migration, and 35
frontend tests green.

+++ backend/src/aerogram/rating/schemas.py
@@ -130,7 +130,10 @@ class RateOfferOut(BaseModel):
     lateness_seconds: int | None = None
     on_time_probability: Decimal | None = Field(default=None, ge=0, le=1)
     probability_label: ProbabilityLabel | None = None
-    carrier_score: Decimal | None = Field(default=None, ge=0, le=100)
+    #: Целое, а не Decimal: колонка ``score_at_quote`` целая, проверка
+    #: ``score_range`` держит 0..100, а Pydantic отдаёт Decimal в JSON строкой —
+    #: и в схеме контракта оказалось бы "92" там, где объявлен number.
+    carrier_score: int | None = Field(default=None, ge=0, le=100)
     risk: RiskLevel | None = None
     confidence: ScoreConfidence | None = None
     eligible: bool
